feat: support .ogg and .mp3
The support for this is a bit more dodgy. FMOD complains when loading _some_ OGG Vorbis or Mpeg files, but others work fine... ?
Like, a .wav file converted into .ogg or .mp3 from Audacity doesn't work, but doing the conversion through ffmpeg works.

Not sure what's going on there, but at least it logs an error so the user knows something's wrong.